requirements.txt
datetime==5.5
fastapi==0.115.7
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic[email]==2.10.6
pytest==8.3.4
uvicorn==0.34.0
passlib[bcrypt]==1.7.4
python-multipart
python-jose[cryptography]